pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.4.0",
    "amqp==5.3.1",
    "asgiref==3.10.0",
    "attrs==25.4.0",
    "autobahn==25.10.2",
    "automat==25.4.16",
    "billiard==4.2.2",
    "celery==5.5.3",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "channels==4.3.1",
    "channels-redis==4.3.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "constantly==23.10.4",
    "cron-descriptor==2.0.6",
    "cryptography==46.0.3",
    "daphne==4.2.1",
    "django==5.2.8",
    "django-celery-beat==2.8.1",
    "django-celery-results==2.6.0",
    "django-cors-headers==4.9.0",
    "django-ratelimit==4.1.0",
    "django-redis==6.0.0",
    "django-timezone-field==7.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-yasg==1.21.11",
    "frozenlist==1.8.0",
    "hyperlink==21.0.0",
    "idna==3.11",
    "incremental==24.7.2",
    "inflection==0.5.1",
    "kombu==5.5.4",
    "msgpack==1.1.2",
    "multidict==6.7.0",
    "packaging==25.0",
    "paho-mqtt==2.1.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pyjwt==2.10.1",
    "pyopenssl==25.3.0",
    "python-crontab==3.3.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis==7.0.1",
    "requests==2.32.5",
    "service-identity==24.2.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "twilio==9.8.6",
    "twisted==25.5.0",
    "txaio==25.9.2",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "vine==5.1.0",
    "wcwidth==0.2.14",
    "yarl==1.22.0",
    "zope-interface==8.1",
]
